monitoring/requirements.txt
functions-framework==3.*
lightgbm==4.*
google-cloud-bigquery==3.*
google-cloud-storage==2.*
pandas==2.*
pyarrow==15.*
db-dtypes==1.*